# apps/api/swagger.json
{"openapi":"3.0.0","paths":{"/v1/auth/google":{"post":{"operationId":"AuthPublicController_loginWithGoogle","parameters":[],"responses":{"200":{"description":"Google login successful","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthGoogleLoginResponseDto"}}}},"401":{"description":"401","content":{"application/json":{"schema":{"oneOf":[{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"auth.error.socialGoogle"},"statusCode":{"type":"number","example":1104}}}]}}}}},"security":[{"bearer":[]},{"google":[]}],"summary":"Google login","tags":["auth.public"]}},"/v1/health":{"get":{"operationId":"AppController_healthCheck","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/v1/user":{"get":{"operationId":"UserController_findAll","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["user"]},"post":{"operationId":"UserController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["user"]}},"/v1/user/{id}":{"get":{"operationId":"UserController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["user"]},"patch":{"operationId":"UserController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["user"]},"delete":{"operationId":"UserController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["user"]}},"/v1/email":{"post":{"operationId":"EmailController_sendEmail","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmailSendDto"}}}},"responses":{"200":{"description":"200","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"email.send"},"statusCode":{"type":"number","example":200}}}}}},"401":{"description":"401","content":{"application/json":{"schema":{"oneOf":[{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"auth.error.accessTokenUnauthorized"},"statusCode":{"type":"number","example":1107}}}]}}}},"408":{"description":"408","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"http.serverError.requestTimeout"},"statusCode":{"type":"number","example":1200}}}}}},"422":{"description":"422","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"request.validation"},"statusCode":{"type":"number","example":1200}}}}}},"500":{"description":"500","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResponseDto"}],"properties":{"message":{"example":"http.serverError.internalServerError"},"statusCode":{"type":"number","example":1000}}}}}}},"security":[{"accessToken":[]}],"summary":"이메일 전송","tags":["modules.email"]}}},"info":{"title":"Sync APIs Specification","description":"Section for describe whole APIs","version":"1.0","contact":{}},"tags":[],"servers":[{"url":"/"}],"components":{"securitySchemes":{"accessToken":{"scheme":"bearer","bearerFormat":"JWT","type":"http"},"google":{"scheme":"bearer","bearerFormat":"JWT","type":"http"},"xApiKey":{"type":"apiKey","in":"header","name":"x-api-key"}},"schemas":{"ResponseMetadataDto":{"type":"object","properties":{}},"ResponseDto":{"type":"object","properties":{"statusCode":{"type":"number","nullable":false,"description":"return specific status code for every endpoints","example":200},"message":{"type":"string","nullable":false,"description":"Message base on language","example":"message endpoint"},"_metadata":{"nullable":false,"description":"Contain metadata about API","example":{"language":"en","timestamp":1660190937231,"timezone":"Asia/Dubai","path":"/api/v1/test/hello","version":"1","repoVersion":"1.0.0"},"allOf":[{"$ref":"#/components/schemas/ResponseMetadataDto"}]}},"required":["statusCode","message","_metadata"]},"AuthGoogleLoginResponseDto":{"type":"object","properties":{"userId":{"type":"number","description":"User id","example":"1"},"accessToken":{"type":"string","description":"Access token","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."},"refreshToken":{"type":"string","description":"Refresh token","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."},"email":{"type":"string","description":"User email","example":"user@example.com"}},"required":["userId","accessToken","refreshToken","email"]},"CreateUserDto":{"type":"object","properties":{}},"UpdateUserDto":{"type":"object","properties":{}},"EmailSendDto":{"type":"object","properties":{"type":{"type":"string","description":"The type of the message","nullable":false,"example":"CHANGE_PASSWORD","enum":["SIGN_UP","CHANGE_PASSWORD","WElCOME","EMAIL"]},"to":{"type":"string","description":"Recipient email address","nullable":false,"example":"recipient@example.com"}},"required":["type","to"]}}}}